requirements.txt
streamlit
pillow==10.2.0